main/requirements.txt
psutil
pystray
Pillow
win10toast